pyproject.toml
[project]
name = "threadwise"
version = "0.1.0"
description = "Project-scoped RAG ingestion for Gmail"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "hatchling>=1.28.0",
    "markdownify>=0.13",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
    "mypy>=1.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/threadwise"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["threadwise"]
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
